+++ Runtime/Core/QosResults.cs
@@ -4,18 +4,18 @@ using Unity.Services.Qos.Internal;
 
 namespace Unity.Services.Qos
 {
-    internal class QosResults : IQosResults
+    class QosResults : IQosResults
     {
-        private IQosService _qosService;
+        WrappedQosService _qosService;
 
-        internal QosResults(IQosService qosService)
+        internal QosResults(WrappedQosService qosService)
         {
             _qosService = qosService;
         }
 
-        public Task<IList<QosResult>> GetSortedQosResultsAsync(string service, IList<string> regions)
+        public Task<IList<Internal.QosResult>> GetSortedQosResultsAsync(string service, IList<string> regions)
         {
-            return _qosService.GetSortedQosResultsAsync(service, regions);
+            return _qosService.GetSortedInternalQosResultsAsync(service, regions);
         }
     }
 }