+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2022-05-27
+
+### Changed
+- Updated dependencies: Core to 1.4.0, Newtonsoft JSON to 3.0.2
+- Fixed an issue making the public interface (QosService.Instance.GetSortedQosResultsAsync(...))
+unusable from the default Unity assembly (Assembly-CSharp)
+
 ## [1.0.0] - 2022-05-04
 
 ### Added